Added JavaDocs for current functons
