+++ Updates.txt
@@ -119,6 +119,12 @@ Datetime does not recognise if user type in single digit dates (3/03/2020)
 
 3/2/20
 A-More OOP
+Added UI class
+
+
+
+A-JavaDocs
+Added Java Docs
