+++ src/main/java/Command.java
@@ -1,5 +1,26 @@
+/**
+ * An interface skeleton for all UserCommands
+ */
 public interface Command {
+    /**
+     * Replies User when command is executed
+     *
+     * @throws DukeException the duke exception
+     */
     public void reply() throws DukeException;
+
+    /**
+     * Executes the Action of the Command
+     *
+     * @return the boolean
+     * @throws DukeException the duke exception
+     */
     public boolean action() throws DukeException;
+
+    /**
+     * Returns the ID of the command
+     *
+     * @return the id
+     */
     public int getID();
 }
