+++ src/main/java/DateTimeHandler.java
@@ -1,3 +1,9 @@
+/**
+ * The class DateTimeHandler acts as a wrapper class for functions to getLocalDate time using only
+ * descriptions
+ * It also handles anything regarding datetime
+ */
+
 import java.time.DateTimeException;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -5,6 +11,13 @@ import java.time.format.DateTimeFormatter;
 import java.util.regex.Pattern;
 
 public class DateTimeHandler{
+    /**
+     * Gets local date time.
+     *
+     * @param desc the desc
+     * @return the local date time
+     * @throws DukeException the duke exception
+     */
     public static LocalDateTime getLocalDateTime(String desc) throws DukeException{
         try {
             if (containsTime(desc)) {
@@ -23,6 +36,12 @@ public class DateTimeHandler{
         }
     }
 
+    /**
+     * Prints local date time string.
+     *
+     * @param dateTime the date time
+     * @return the string
+     */
     public static String printsLocalDateTime(LocalDateTime dateTime){
         if(dateTime == null){
             return "";
@@ -35,7 +54,12 @@ public class DateTimeHandler{
         }
     }
 
-    //understands 2/12/2019 1800
+    /**
+     * Checks if given string contains time format
+     *
+     * @param desc description
+     * @return boolean value if description contains time syntax
+     */
     private static boolean containsTime(String desc){
         String pattern = "[0-9]{2}/[0-9]{2}/[0-9]{4} [0-9]{4}";
         String patternEng = "[0-9]{2} [a-zA-Z]{3} [0-9]{4} [0-9]{2}:[0-9]{2}[apmAPM]{2}";
@@ -43,17 +67,35 @@ public class DateTimeHandler{
         return Pattern.matches(pattern, desc)||Pattern.matches(patternEng, desc) || Pattern.matches(patternEng2, desc);
     }
 
+    /**
+     * Checks if given string is valid date
+     *
+     * @param date date in string
+     * @return boolean value if string is a valid date
+     */
     private static boolean isValidDate(String date){
         String pattern = "[0-9]{2}/[0-9]{2}/[0-9]{4}";
         return Pattern.matches(pattern, date);
     }
 
+
+    /**
+     * Checks if given string is valid time
+     *
+     * @param time time in string
+     * @return boolean value if string is a valid time
+     */
     private static boolean isValidTime(String time){
         String pattern = "[0-9]{4}";
         return Pattern.matches(pattern, time);
     }
 
-
+    /**
+     * Reformats the dateTime
+     *
+     * @param dateTime
+     * @return String of the dateTime in correct format
+     */
     private static String printDateAndTime(LocalDateTime dateTime){
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("d MMM yyyy hh:mma");
         return dateTime.format(formatter);
@@ -69,6 +111,12 @@ public class DateTimeHandler{
         return dateTime.format(formatter);
     }
 
+    /**
+     * Reverts date time string for easier parse to memory.
+     *
+     * @param dateTime the date time
+     * @return the string for memory storage
+     */
     public static String revertDateTime(String dateTime){
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("d MMM yyyy hh:mma");;
         DateTimeFormatter formatterSave = DateTimeFormatter.ofPattern("dd/MM/yyyy hhmm");
