--- src/main/java/Del_Command.java
@@ -1,27 +1,0 @@
-public class Del_Command extends UserCommand {
-
-    private static int id;
-    Del_Command(int id){
-        super(5);
-        setDel(id);
-    }
-
-    @Override
-    public void reply(){
-        System.out.println(UI.getReply("delete"));
-    }
-
-
-    public static void setDel(int ID){
-        id = ID;
-    }
-
-    @Override
-    public boolean action() throws DukeException {
-
-        MyList.deleteItem(id);
-        MyList.printCount();
-        return true;
-    }
-
-}
