+++ src/main/java/DukeException.java
@@ -1,4 +1,12 @@
+/**
+ * The type Duke exception.
+ */
 public class DukeException extends Exception {
+    /**
+     * Instantiates a new Duke exception.
+     *
+     * @param message the message
+     */
     DukeException(String message){
         super(message);
     }
