+++ src/main/java/ListCommand.java
@@ -1,6 +1,10 @@
-import java.util.HashMap;
-
+/**
+ * The command handles when user calls list
+ */
 public class ListCommand extends UserCommand {
+    /**
+     * Instantiates a new List command.
+     */
     ListCommand() {
         super(3);
     }
