+++ src/main/java/MyList.java
@@ -1,14 +1,28 @@
 import java.util.ArrayList;
 
+/**
+ * MyList handles everything to do with the list
+ */
 public class MyList {
     private static ArrayList<Task> DaList = new ArrayList<>();
 
 
+    /**
+     * Add tasks.
+     *
+     * @param t the t
+     */
     public static void addItem(Task t){
         System.out.println("added: " + t);
         DaList.add(t);
     }
 
+    /**
+     * Delete tasks.
+     *
+     * @param ID the id
+     * @throws DukeException the duke exception
+     */
     public static void deleteItem(int ID)throws DukeException{
         ID = ID-1;
         if(ID>=DaList.size()|| ID<0){
@@ -19,6 +33,9 @@ public class MyList {
     }
 
 
+    /**
+     * Prints list.
+     */
     public static void printList(){
         if(DaList.size()==0){
             System.out.println(UI.getReply("emptyList"));
@@ -30,10 +47,19 @@ public class MyList {
         }
     }
 
+    /**
+     * Prints number of items in tasks.
+     */
     public static void printCount(){
         System.out.println("Now you have "+ DaList.size() + " things in your bloody list..");
     }
 
+    /**
+     * Mark list item as done.
+     *
+     * @param ID the id
+     * @throws DukeException the duke exception
+     */
     public static void markDone(int ID) throws DukeException{
         ID = ID-1;
         if(ID>=DaList.size()|| ID<0){
@@ -43,6 +69,11 @@ public class MyList {
         System.out.println(DaList.get(ID));
     }
 
+    /**
+     * Get list array list.
+     *
+     * @return the list of objects
+     */
     public static ArrayList getList(){
         return DaList;
     }
