+++ src/main/java/Task.java
@@ -1,22 +1,46 @@
+/**
+ * The type Task.
+ */
 public class Task {
 
     private String Description = "";
     protected boolean isDone = false;
 
+    /**
+     * Instantiates a new Task.
+     *
+     * @param description the description
+     */
     public Task(String description) {
         this.Description = description;
     }
 
+    /**
+     * Reset desc.
+     *
+     * @param desc the desc
+     */
     protected void resetDesc(String desc){
         Description = desc;
     }
 
+    /**
+     * Mark done.
+     *
+     * @throws DukeException the duke exception
+     */
     public void markDone() throws DukeException{
         if(isDone){
             throw new DukeException(UI.getReply("doneAlready"));
         }
         isDone = true;
     }
+
+    /**
+     * Gets status icon.
+     *
+     * @return the status icon
+     */
     public String getStatusIcon() {
         return (isDone ? "\u2713" : "\u2718"); //return tick or X symbols
     }
