+++ src/main/java/ToDos.java
@@ -1,4 +1,12 @@
+/**
+ * ToDo is a type of task that accepts /at statements only
+ */
 public class ToDos extends Task {
+    /**
+     * Instantiates a new To dos.
+     *
+     * @param description the description
+     */
     public ToDos(String description) {
         super(description);
     }
